+++ st2actions/st2actions/runners/mistral/v2.py
@@ -53,7 +53,9 @@ class MistralRunner(AsyncActionRunner):
             username=cfg.CONF.mistral.keystone_username,
             api_key=cfg.CONF.mistral.keystone_password,
             project_name=cfg.CONF.mistral.keystone_project_name,
-            auth_url=cfg.CONF.mistral.keystone_auth_url)
+            auth_url=cfg.CONF.mistral.keystone_auth_url,
+            cacert=cfg.CONF.mistral.cacert,
+            insecure=cfg.CONF.mistral.insecure)
 
     def pre_run(self):
         if getattr(self, 'liveaction', None):