+++ st2api/st2api/cmd/api.py
@@ -24,9 +24,6 @@ from st2common import log as logging
 from st2common.service_setup import setup as common_setup
 from st2common.service_setup import teardown as common_teardown
 from st2common.util.monkey_patch import monkey_patch
-from st2common.util.wsgi import shutdown_server_kill_pending_requests
-from st2api.signal_handlers import register_api_signal_handlers
-from st2api.listener import get_listener_if_set
 from st2api import config
 config.register_opts()
 from st2api import app
@@ -58,15 +55,6 @@ def _run_server():
     worker_pool = eventlet.GreenPool(max_pool_size)
     sock = eventlet.listen((host, port))
 
-    def queue_shutdown(signal_number, stack_frame):
-        eventlet.spawn_n(shutdown_server_kill_pending_requests, sock=sock,
-                         worker_pool=worker_pool, wait_time=WSGI_SERVER_REQUEST_SHUTDOWN_TIME)
-
-    # We register a custom SIGINT handler which allows us to kill long running active requests.
-    # Note: Eventually we will support draining (waiting for short-running requests), but we
-    # will still want to kill long running stream requests.
-    register_api_signal_handlers(handler_func=queue_shutdown)
-
     wsgi.server(sock, app.setup_app(), custom_pool=worker_pool)
     return 0
 
@@ -81,11 +69,6 @@ def main():
         return _run_server()
     except SystemExit as exit_code:
         sys.exit(exit_code)
-    except KeyboardInterrupt:
-        listener = get_listener_if_set()
-
-        if listener:
-            listener.shutdown()
     except Exception:
         LOG.exception('(PID=%s) ST2 API quit due to exception.', os.getpid())
         return 1