+++ st2api/st2api/controllers/v1/root.py
@@ -26,12 +26,15 @@ from st2api.controllers.v1.rules import RuleController
 from st2api.controllers.v1.rule_enforcements import RuleEnforcementController
 from st2api.controllers.v1.runnertypes import RunnerTypesController
 from st2api.controllers.v1.sensors import SensorTypeController
-from st2api.controllers.v1.stream import StreamController
 from st2api.controllers.v1.traces import TracesController
 from st2api.controllers.v1.triggers import TriggerTypeController, TriggerController, \
     TriggerInstanceController
 from st2api.controllers.v1.webhooks import WebhooksController
 from st2api.controllers.v1.rbac import RBACController
+
+__all__ = [
+    'RootController'
+]
 
 
 class RootController(object):
@@ -51,7 +54,6 @@ class RootController(object):
     ruleenforcements = RuleEnforcementController()
     keys = KeyValuePairController()
     webhooks = WebhooksController()
-    stream = StreamController()
     actionalias = ActionAliasController()
     aliasexecution = ActionAliasExecutionController()
     traces = TracesController()