+++ st2common/st2common/config.py
@@ -136,7 +136,11 @@ def register_opts(ignore_errors=False):
     # Common API options
     api_opts = [
         cfg.StrOpt('host', default='0.0.0.0', help='StackStorm API server host'),
-        cfg.IntOpt('port', default=9101, help='StackStorm API server port')
+        cfg.IntOpt('port', default=9101, help='StackStorm API server port'),
+        cfg.ListOpt('allow_origin', default=['http://127.0.0.1:3000'],
+                    help='List of origins allowed for api, auth and stream'),
+        cfg.BoolOpt('mask_secrets', default=True,
+                    help='True to mask secrets in the API responses')
     ]
     do_register_opts(api_opts, 'api', ignore_errors)
 
@@ -171,6 +175,8 @@ def register_opts(ignore_errors=False):
         cfg.StrOpt('keystone_password', default=None, help='Password for authentication.'),
         cfg.StrOpt('keystone_project_name', default=None, help='OpenStack project scope.'),
         cfg.StrOpt('keystone_auth_url', default=None, help='Auth endpoint for Keystone.'),
+        cfg.StrOpt('cacert', default=None, help='Optional certificate to validate endpoint.'),
+        cfg.BoolOpt('insecure', default=False, help='Allow insecure communication with Mistral.'),
 
         cfg.StrOpt('api_url', default=None, help=('URL Mistral uses to talk back to the API.'
             'If not provided it defaults to public API URL. Note: This needs to be a base '