+++ st2common/st2common/triggers.py
@@ -16,6 +16,7 @@
 import six
 
 from mongoengine import ValidationError
+from mongoengine import NotUniqueError
 from oslo_config import cfg
 
 from st2common import log as logging
@@ -31,7 +32,13 @@ LOG = logging.getLogger(__name__)
 
 
 def _register_internal_trigger_type(trigger_definition):
-    trigger_type_db = create_trigger_type_db(trigger_type=trigger_definition)
+    try:
+        trigger_type_db = create_trigger_type_db(trigger_type=trigger_definition)
+    except (NotUniqueError, StackStormDBObjectConflictError):
+        # We ignore conflict error since this operation is idempodent and race is not an
+        # issue
+        LOG.debug('Trigger type "%s" already exists, ignoring...' %
+                  (trigger_definition['name']), exc_info=True)
 
     # trigger types with parameters do no require a shadow trigger.
     if not trigger_type_db.parameters_schema:
@@ -42,13 +49,14 @@ def _register_internal_trigger_type(trigger_definition):
             extra = {'trigger_db': trigger_db}
             LOG.audit('Trigger created for parameter-less TriggerType. Trigger.id=%s' %
                       (trigger_db.id), extra=extra)
+        except StackStormDBObjectConflictError:
+            LOG.debug('Shadow trigger "%s" already exists. Ignoring.',
+                      trigger_type_db.get_reference().ref, exc_info=True)
+
         except (ValidationError, ValueError):
             LOG.exception('Validation failed in shadow trigger. TriggerType=%s.',
                           trigger_type_db.get_reference().ref)
             raise
-        except StackStormDBObjectConflictError:
-            LOG.debug('Shadow trigger "%s" already exists. Ignoring.',
-                      trigger_type_db.get_reference().ref, exc_info=True)
 
     return trigger_type_db
 
@@ -73,7 +81,7 @@ def register_internal_trigger_types():
             try:
                 trigger_type_db = _register_internal_trigger_type(
                     trigger_definition=trigger_definition)
-            except:
+            except Exception:
                 LOG.exception('Failed registering internal trigger: %s.', trigger_definition)
                 raise
             else: